W257: the `slti` destination is a source-shape defect — +7 MATCH
**6623 -> 6630 first-party MATCH (80.5%).** 0 lost, 0 new over-window, 0 regressions.

## The new lever: `slti $at` versus `slti $v0`

Several functions sat at nd2 with a single residual — the destination register of a comparison and its
dependent branch. It had been recorded as an allocator floor for the whole campaign. It is not.

**MWCC emits `slti $at` when the comparison is a throwaway feeding an immediate branch, and allocates a
normal register when the result is a named value with a lifetime.** So the register tells you whether
the original source named the comparison.

Three MATCHes, and the lever works in **both** directions:

- **`fclShopMisc 003fee40` nd2 -> 0.** Retail used `$v0`, so the comparison was named. `u32 ok = (value
  < 0x63); if (ok == 0)` produced retail's `slti $v0` and `bnez $v0` exactly. Note the type and the test
  polarity both mattered: `s32 ok` gave nd3, `u32 ok` with `if (ok)` gave nd1 but inverted the
  semantics, and only `u32` with `if (ok == 0)` was both correct and nd0.
- **`fcl_misc 003c9850` nd2 -> 0.** Preserving retail's separate `bltz` guard while naming only the
  `slti`-fed comparison allocated `$v0` without disturbing the layout.
- **`mt_evtEffect 00395e20` nd2 -> 0, the opposite direction.** Here *ours* used a named register and
  retail used `$at`, so the fix was to stop naming it: a clean `for` condition `iVar6 <= 0x28` made MWCC
  use retail's throwaway comparison.

`mt_scene 003b5620` nd19 -> 15 by applying it at four range clauses.

## Four more MATCHes from the established levers

- **`mdlFile 0034c1b0` nd2 -> 0.** Two independent pointer `addiu` instructions were reversed before a
  memcpy-like call whose declaration was already correct. Naming the destination and source pointers as
  locals reversed them into retail's order. **So an argument-order residual can be a
  value-formation-lifetime defect even when the callee signature is right** — a genuinely new variant.
- **`y_runtime 004560d0` nd10 -> 0.** Reordering the mixed-ABI parameters at source level while
  preserving the physical GPR/FPR assignments made the prologue preserve `$a1`, then `$a2`, then `$f14`
  exactly as retail. `004344f0` nd13 -> 5 by the same mechanism.
- **`y_runtime 0045b4e0` nd1 -> 0.** Introducing a meaningful `tableOffset` local changed the sole
  commutative `addu` from ours `$v0,$a0,$v0` to retail `$v0,$v0,$a0`.
- **`op_menu 0026cc90`** reached nd0, completing a function an earlier wave had brought from nd1089 to 12.

Also `btlFormation 002c3ce0` nd6 -> 2 by correcting `score` from `u8` to `u16`, which removed a
retail-absent `andi`.

## An important qualification, from a disagreement between two agents

W257Itf ran a careful, well-designed test of whether commutative `addu` orientation at nd1 is terminal:
five functions, multiple naming and lifetime and declaration-order probes each, and **every probe failed
or regressed** (`itfMesManager 003a4a70` went nd1 -> 90). It concluded terminal.

W257Yaj then MATCHed `y_runtime 0045b4e0` — nd1, commutative `addu` — by naming an operand.

**Both are right: operand orientation follows allocation order, which follows value lifetime, so it is
reachable in principle but only when the source genuinely has a nameable intermediate.** Where the two
operands are already minimal expressions there is nothing to name and it is terminal. That is a much
more useful rule than either verdict alone, and it means an nd1 commutative `addu` deserves exactly one
naming probe, not a campaign.

The nd<=20 band went 140 -> 133. Remaining residuals in it are now overwhelmingly cyclic saved-register
colouring with matching frames, `addiu`-versus-`daddiu` where retail genuinely uses 64-bit arithmetic
(`btlBoss 002f88c0`, five probes all reverted), and independent non-argument instruction scheduling —
each now recorded with the specific probes that failed.

+++ src/Battle/btlFormation.c
@@ -7430,7 +7430,7 @@ u32 func_002c3ce0(u32 param_1)
   int selected;
   u16 index;
   u16 best;
-  u8 score;
+  u16 score;
   extern u8 func_002ffcc0(u32);
 
   result = func_002bff60_u16(param_1,param_1 + 0x88,*(u16 *)(param_1 + 0x6e),0);
